requirements.txt
streamlit==1.52.1
opencv-python-headless==4.12.0.88
pillow==12.0.0
numpy==2.2.6
ultralytics==8.3.235